# [PERF] stock: do optimal scan & join warehouse_cte