@robodoo 2025-03-12
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/201318.png)](https://mergebot.odoo.com/odoo/odoo/pull/201318)