@ethanrobv 2025-03-12
[analyze_explain_old_query.json](https://github.com/user-attachments/files/19208268/analyze_explain_old_query.json)
[Old Query Visualized](https://explain.dalibo.com/plan/39ab7eg32f33fb66)

[analyze_explain_new_query.json](https://github.com/user-attachments/files/19208271/analyze_explain_new_query.json)
[New Query Visualized](https://explain.dalibo.com/plan/4ge812f485f5b6d8)

doing the join on an expression has the same issue that the original query did- namely it does not facilitate good usage of indexes- in this case you can see it means the scan of the wareouse CTE has a greatly inflated number of rows, which means both that node itself is slow and also later operations have poorer performance.

The proposed change in this PR's commit uses a valued case expression (i.e., `JOIN cte ON cte.col = CASE...`) which does not impeded use of indexes, and does result in the inflated number of rows during the warehouse cte scan.

|                | Old Query | New Query |
|----------------|-----------|-----------|
| Execution Time | 1m 18s    | 9s 271 ms |